moved shared traits
